Implement GroupProcessor and update schema processing interfaces for enhanced handling of group entries
