+++ src/common/services/recordGenerator/processors/schema/schemaProcessor.interface.ts
@@ -1,5 +1,5 @@
 export interface ISchemaProcessor {
   canProcess(schemaEntry: SchemaEntry, modelField: RecordModelField): boolean;
 
-  process(schemaEntry: SchemaEntry, userValues: UserValues): Record<string, any>;
+  process(schemaEntry: SchemaEntry, userValues: UserValues, modelField?: RecordModelField): Record<string, any>;
 }
