+++ src/common/services/recordGenerator/processors/schema/schemaProcessorManager.ts
@@ -1,13 +1,18 @@
 import { SchemaManager } from '../../schemaManager';
-import { DropdownProcessor } from './dropdownProcessor';
 import { ISchemaProcessor } from './schemaProcessor.interface';
+import { DropdownProcessor } from './dropdownProcessor';
+import { GroupProcessor } from './GroupProcessor';
 import { UnwrappedDropdownOptionProcessor } from './unwrappedDropdownOptionProcessor';
 
 export class SchemaProcessorManager {
   private readonly processors: ISchemaProcessor[];
 
   constructor(schemaManager: SchemaManager) {
-    this.processors = [new UnwrappedDropdownOptionProcessor(schemaManager), new DropdownProcessor(schemaManager)];
+    this.processors = [
+      new GroupProcessor(schemaManager),
+      new DropdownProcessor(schemaManager),
+      new UnwrappedDropdownOptionProcessor(schemaManager),
+    ];
   }
 
   registerProcessor(processor: ISchemaProcessor) {
@@ -17,7 +22,7 @@ export class SchemaProcessorManager {
   process(schemaEntry: SchemaEntry, modelField: RecordModelField, userValues: UserValues): Record<string, any> {
     for (const processor of this.processors) {
       if (processor.canProcess(schemaEntry, modelField)) {
-        return processor.process(schemaEntry, userValues);
+        return processor.process(schemaEntry, userValues, modelField);
       }
     }
