+++ src/common/services/recordGenerator/recordGenerator.ts
@@ -3,26 +3,7 @@ import { RecordModelType } from '@common/constants/recordModel.constants';
 import { IRecordGenerator } from './recordGenerator.interface';
 import { SchemaManager } from './schemaManager';
 import { SchemaProcessorManager } from './processors/schema/schemaProcessorManager';
-import { GeneratedValue, ValueOptions, ValueResult } from './processors/schema/valueTypes';
-
-interface UserValueContent {
-  id?: string;
-  label: string;
-  meta?: {
-    uri?: string;
-    basicLabel?: string;
-  };
-}
-
-interface ChildEntryWithValues {
-  childEntry: SchemaEntry;
-  childValues: UserValueContent[];
-}
-
-type GroupedValue = Array<{
-  childEntry: SchemaEntry;
-  valueAtIndex: UserValueContent | null;
-}>;
+import { GeneratedValue, UserValueContent, ValueOptions, ValueResult } from './types/valueTypes';
 
 export class RecordGenerator implements IRecordGenerator {
   private readonly schemaManager: SchemaManager;
@@ -69,12 +50,13 @@ export class RecordGenerator implements IRecordGenerator {
 
   private generateValueFromModel(modelField: RecordModelField, schemaEntry: SchemaEntry) {
     const options: ValueOptions = {};
-
     const processorValue = this.schemaProcessorManager.process(schemaEntry, modelField, this.userValues);
+
     if (Object.keys(processorValue).length > 0) {
       if (modelField.options?.hiddenWrapper) {
         options.hiddenWrapper = true;
       }
+
       return { value: processorValue, options };
     }
 
@@ -83,6 +65,7 @@ export class RecordGenerator implements IRecordGenerator {
 
     if (modelField.type === RecordModelType.array) {
       const arrayResult = this.processArrayType(modelField, schemaEntry, values as UserValueContent[]);
+
       return { value: arrayResult, options };
     }
 
@@ -110,12 +93,10 @@ export class RecordGenerator implements IRecordGenerator {
     }
 
     // Handle regular groups with children (e.g. accessLocation, supplementaryContent, _notes)
-    if (
-      modelField.value === RecordModelType.object &&
-      schemaEntry.children &&
-      schemaEntry.type === AdvancedFieldType.group
-    ) {
-      return this.processGroupWithChildren(modelField, schemaEntry);
+    const processorValue = this.schemaProcessorManager.process(schemaEntry, modelField, this.userValues);
+
+    if (Object.keys(processorValue).length > 0) {
+      return processorValue;
     }
 
     return null;
@@ -176,6 +157,7 @@ export class RecordGenerator implements IRecordGenerator {
   private processArrayField(key: string, childResult: ValueResult, result: GeneratedValue) {
     if (childResult.options.hiddenWrapper) {
       const firstValue = Array.isArray(childResult.value) ? childResult.value[0] : null;
+
       if (firstValue && typeof firstValue === 'object') {
         Object.assign(result, firstValue);
       }
@@ -192,111 +174,5 @@ export class RecordGenerator implements IRecordGenerator {
         }
       }
     }
-  }
-
-  private processGroupWithChildren(modelField: RecordModelField, groupEntry: SchemaEntry) {
-    if (!groupEntry.children || !modelField.fields) {
-      return [];
-    }
-
-    const childEntriesWithValues = this.getChildEntriesWithValues(groupEntry.children);
-
-    if (childEntriesWithValues.length === 0) {
-      return [];
-    }
-
-    const groupedValues = this.groupValuesByIndex(childEntriesWithValues);
-
-    return this.createStructuredObjects(groupedValues, modelField.fields);
-  }
-
-  private getChildEntriesWithValues(children: string[]) {
-    return children
-      .map(childUuid => {
-        const childEntry = this.schemaManager.getSchemaEntry(childUuid);
-
-        if (!childEntry?.type || !childEntry.uriBFLite) return null;
-
-        const childValues = this.userValues[childEntry.uuid]?.contents || [];
-
-        if (childValues.length === 0) return null;
-
-        return { childEntry, childValues };
-      })
-      .filter((entry): entry is ChildEntryWithValues => entry !== null);
-  }
-
-  private createStructuredObjects(groupedValues: GroupedValue[], modelFields: Record<string, RecordModelField>) {
-    return groupedValues
-      .map(indexGroup => this.createGroupObject(indexGroup, modelFields))
-      .filter(obj => Object.keys(obj).length > 0);
-  }
-
-  private createGroupObject(indexGroup: GroupedValue, modelFields: Record<string, RecordModelField>): GeneratedValue {
-    const groupObject: GeneratedValue = {};
-
-    for (const { childEntry, valueAtIndex } of indexGroup) {
-      if (!childEntry.uriBFLite || !valueAtIndex) continue;
-
-      this.mapValueToModelField(childEntry, valueAtIndex, modelFields, groupObject);
-    }
-
-    return groupObject;
-  }
-
-  private mapValueToModelField(
-    childEntry: SchemaEntry,
-    valueAtIndex: UserValueContent,
-    modelFields: Record<string, RecordModelField>,
-    groupObject: GeneratedValue,
-  ) {
-    const { uriBFLite, type } = childEntry;
-
-    if (!uriBFLite || type === undefined) return;
-
-    const fieldType = this.validateFieldType(type);
-
-    if (fieldType === null) return;
-
-    for (const key of Object.keys(modelFields)) {
-      if (key === uriBFLite) {
-        groupObject[key] = this.getValueForType(fieldType, valueAtIndex);
-      }
-    }
-  }
-
-  private validateFieldType(type: string) {
-    return Object.values(AdvancedFieldType).includes(type as AdvancedFieldType) ? (type as AdvancedFieldType) : null;
-  }
-
-  private getValueForType(type: AdvancedFieldType, value: UserValueContent) {
-    switch (type) {
-      case AdvancedFieldType.literal:
-        return [value.label];
-      case AdvancedFieldType.simple:
-        return [value.meta?.uri ?? value.label];
-      default:
-        return [];
-    }
-  }
-
-  private groupValuesByIndex(childEntriesWithValues: ChildEntryWithValues[]) {
-    const maxValueCount = this.getMaxValueCount(childEntriesWithValues);
-    return Array.from({ length: maxValueCount }, (_, index) =>
-      this.createValueGroup(childEntriesWithValues, index),
-    ).filter(group => group.length > 0);
-  }
-
-  private getMaxValueCount(childEntriesWithValues: ChildEntryWithValues[]) {
-    return Math.max(...childEntriesWithValues.map(({ childValues }) => childValues.length), 0);
-  }
-
-  private createValueGroup(childEntriesWithValues: ChildEntryWithValues[], index: number) {
-    return childEntriesWithValues
-      .map(({ childEntry, childValues }) => ({
-        childEntry,
-        valueAtIndex: childValues[index] || null,
-      }))
-      .filter(({ valueAtIndex }) => valueAtIndex !== null);
   }
 }
